itimorblog/requirements.txt
Django==1.11.2
docutils==0.13.1
olefile==0.44
Pillow==4.2.0
django-pagedown==0.1.3
django-uuslug==1.1.8
Pygments==2.2.0
markwon==2.3.4